+++ luatest/runner.lua
@@ -4,6 +4,7 @@
 
 local clock = require('clock')
 
+local assertions = require('luatest.assertions')
 local capturing = require('luatest.capturing')
 local Class = require('luatest.class')
 local GenericOutput = require('luatest.output.generic')
@@ -274,6 +275,8 @@ function Runner.mt:start_suite(selected_count, not_selected_count)
             fail = {},
             error = {},
             skip = {},
+            xfail = {},
+            xsuccess = {},
         },
     }
     self.output.result = self.result
@@ -298,7 +301,8 @@ function Runner.mt:update_status(node, err)
     -- if the node is already in failure/error, just don't report the new error (see above)
     elseif not node:is('success') then
         return
-    elseif err.status == 'fail' or err.status == 'error' or err.status == 'skip' then
+    elseif err.status == 'fail' or err.status == 'error' or err.status == 'skip'
+        or err.status == 'xfail' or err.status == 'xsuccess' then
         node:update_status(err.status, err.message, err.trace)
     else
         error('No such status: ' .. pp.tostring(err.status))
@@ -311,9 +315,10 @@ function Runner.mt:end_test(node)
     node.start_time = nil
     self.output:end_test(node)
 
-    if node:is('error') or node:is('fail') then
+    if node:is('error') or node:is('fail') or node:is('xsuccess') then
         self.result.aborted = self.fail_fast
-    elseif not node:is('success') and not node:is('skip') then
+    elseif not node:is('success') and not node:is('skip')
+        and not node:is('xfail') then
         error('No such node status: ' .. pp.tostring(node.status))
     end
 end
@@ -330,7 +335,7 @@ function Runner.mt:end_suite()
     for _, test in pairs(self.result.tests.all) do
         table.insert(self.result.tests[test.status], test)
     end
-    self.result.failures_count = #self.result.tests.fail + #self.result.tests.error
+    self.result.failures_count = #self.result.tests.fail + #self.result.tests.error + #self.result.tests.xsuccess
     self.output:end_suite()
 end
 
@@ -348,12 +353,27 @@ function Runner.mt:protected_call(instance, method, pretty_name)
         end
     end)
 
+    -- check if test was marked as xfail and reset xfail flag
+    local xfail = assertions.private.is_xfail()
+
     if type(err.message) ~= 'string' then
         err.message = pp.tostring(err.message)
     end
 
-    if err.status == 'success' or err.status == 'skip' then
+    if (err.status == 'success' and not xfail) or err.status == 'skip' then
         err.trace = nil
+        return err
+    end
+
+    if xfail and err.status ~= 'error' and err.status ~= 'skip' then
+        err.status = 'x' .. err.status
+
+        if err.status == 'xsuccess' then
+            err.trace = ''
+            err.message = type(xfail) == 'string' and xfail
+            or 'Test expected to fail has succeeded. Consider removing xfail.'
+        end
+
         return err
     end